LayoutTests/imported/w3c/web-platform-tests/trusted-types/GlobalEventHandlers-onclick-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


PASS a.setAttribte('onclick') sets a trusted script.
FAIL a.setAttribute('onclick') sets an unsuitable trusted type. assert_unreached: Reached unreachable code
FAIL a.setAttribute('click') sets a test string. assert_unreached: Reached unreachable code
